+++ database/migrations/update/0_1/2026_03_24_094948_update_clinical_impressions_table.php
@@ -14,11 +14,21 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('clinical_impressions', function (Blueprint $table) {
-            $table->foreignId('person_id')->after('uuid')->constrained('persons');
-            $table->string('explanatory_letter')->after('note')->nullable();
-            $table->timestamp('ehealth_inserted_at')->nullable()->after('note');
-            $table->timestamp('ehealth_updated_at')->nullable()->after('ehealth_inserted_at');
-            $table->foreignId('encounter_internal_id')->nullable()->change();
+            if (!Schema::hasColumn('clinical_impressions', 'person_id')) {
+                $table->foreignId('person_id')->after('uuid')->constrained('persons');
+            }
+            if (!Schema::hasColumn('clinical_impressions', 'explanatory_letter')) {
+                $table->string('explanatory_letter')->after('note')->nullable();
+            }
+            if (!Schema::hasColumn('clinical_impressions', 'ehealth_inserted_at')) {
+                $table->timestamp('ehealth_inserted_at')->nullable()->after('note');
+            }
+            if (!Schema::hasColumn('clinical_impressions', 'ehealth_updated_at')) {
+                $table->timestamp('ehealth_updated_at')->nullable()->after('ehealth_inserted_at');
+            }
+            if (Schema::hasColumn('clinical_impressions', 'encounter_internal_id')) {
+                $table->foreignId('encounter_internal_id')->nullable()->change();
+            }
         });
     }
 
@@ -28,10 +38,22 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('clinical_impressions', function (Blueprint $table) {
-            $table->foreignId('encounter_internal_id')->nullable(false)->change();
-            $table->dropColumn(['explanatory_letter', 'ehealth_updated_at', 'ehealth_inserted_at']);
-            $table->dropForeign(['person_id']);
-            $table->dropColumn('person_id');
+            if (Schema::hasColumn('clinical_impressions', 'encounter_internal_id')) {
+                $table->foreignId('encounter_internal_id')->nullable(false)->change();
+            }
+            if (Schema::hasColumn('clinical_impressions', 'explanatory_letter')) {
+                $table->dropColumn('explanatory_letter');
+            }
+            if (Schema::hasColumn('clinical_impressions', 'ehealth_updated_at')) {
+                $table->dropColumn('ehealth_updated_at');
+            }
+            if (Schema::hasColumn('clinical_impressions', 'ehealth_inserted_at')) {
+                $table->dropColumn('ehealth_inserted_at');
+            }
+            if (Schema::hasColumn('clinical_impressions', 'person_id')) {
+                $table->dropForeign(['person_id']);
+                $table->dropColumn('person_id');
+            }
         });
     }
 };
